refactor: security 관련 패키지 구조 통합
