+++ src/main/java/io/github/seonrizee/scheduler/domain/comment/service/CommentServiceImpl.java
@@ -9,7 +9,7 @@ import io.github.seonrizee.scheduler.domain.comment.repository.CommentRepository
 import io.github.seonrizee.scheduler.domain.schedule.entity.Schedule;
 import io.github.seonrizee.scheduler.domain.schedule.service.ScheduleFinder;
 import io.github.seonrizee.scheduler.domain.user.entity.User;
-import io.github.seonrizee.scheduler.security.service.AuthorizationService;
+import io.github.seonrizee.scheduler.global.security.service.AuthorizationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
