+++ src/main/java/io/github/seonrizee/scheduler/domain/user/service/UserServiceImpl.java
@@ -6,7 +6,7 @@ import io.github.seonrizee.scheduler.domain.user.dto.response.UserProfileRespons
 import io.github.seonrizee.scheduler.domain.user.entity.User;
 import io.github.seonrizee.scheduler.domain.user.mapper.UserMapper;
 import io.github.seonrizee.scheduler.domain.user.repository.UserRepository;
-import io.github.seonrizee.scheduler.global.config.security.PasswordEncoderConfig;
+import io.github.seonrizee.scheduler.global.security.config.PasswordEncoderConfig;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
